project/app/requirements.txt
certifi==2018.1.18
chardet==3.0.4
dj-database-url==0.5.0
Django==2.0.3
django-cors-headers==2.2.0
django-filter==1.1.0
djangorestframework==3.8.0
djangorestframework-word-filter==0.1.1
idna==2.6
numpy==1.14.2
pandas==0.22.0
pipenv==11.9.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
python-dateutil==2.7.0
pytz==2018.3
requests==2.18.4
six==1.11.0
urllib3==1.22
virtualenv==15.2.0
virtualenv-clone==0.3.0
